third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-preline-eol-expected.txt

This is a testharness.js-based test.
FAIL TEXTAREA with white-space:pre-line should not produce nbsp by Enter key assert_equals: expected "P1\n\nP2\n\n\n\nP3\n" but got "P1\n\nP2\n\n \nP3\n"
Harness: the test ran to completion.
